Fixes issues with huge texts
std::size_t used instead of int for size !
 => lots of implication, + secus & optim (const or const&) .

+ add some missing std:: before pre-used size_t.

Nota src/math/func.h : adding a template instead of 3 func and adding a similare one for std::size_t...

+++ src/CBot/CBotInstr/CBotExprVar.cpp
@@ -213,10 +213,10 @@ CBotInstr* CBotExprVar::CompileMethode(CBotToken* &p, CBotCStack* pStack)
                 if (pStk->IsOk())
                 {
                     inst->AddNext3(i);                            // add after
-                    p = pp;                                        // previous instruction
+                    p = pp;                                       // previous instruction
                     return pStack->Return(inst, pStk);
                 }
-                pStk->SetError(CBotNoErr, 0);                            // the error is not adressed here
+                pStk->SetError(CBotNoErr, static_cast<std::size_t>(0));             // the error is not adressed here
             }
         }
         delete inst;
